@W-14702830@: Refactor part 7 - Moving all business logic out of commands and into Action classes
