+++ src/commands/scanner/rule/add.ts
@@ -1,23 +1,24 @@
 import {Flags} from '@salesforce/sf-plugins-core';
-import {AnyJson} from '@salesforce/ts-types';
-import {Controller} from '../../../Controller';
-import {ScannerCommand} from '../../../lib/ScannerCommand';
+import {Action, ScannerCommand} from '../../../lib/ScannerCommand';
 import {Bundle, getMessage} from "../../../MessageCatalog";
-import {Inputs} from "../../../types";
-import {Config} from "@oclif/core";
-import {PathResolver, PathResolverImpl} from "../../../lib/PathResolver";
+import {PathResolverImpl} from "../../../lib/PathResolver";
+import {Display} from "../../../lib/Display";
+import {RuleAddAction} from "../../../lib/actions/RuleAddAction";
+import {Logger} from "@salesforce/core";
 
-import {InputValidatorFactory, RuleAddCommandInputValidatorFactory} from "../../../lib/InputValidator";
-
+/**
+ * Defines the "rule add" command for the "scanner" cli.
+ */
 export default class Add extends ScannerCommand {
-
+	// These determine what's displayed when the --help/-h flag is provided.
 	public static summary = getMessage(Bundle.Add, 'commandSummary');
 	public static description = getMessage(Bundle.Add, 'commandDescription');
-
 	public static examples = [
 		getMessage(Bundle.Add, 'examples')
 	];
 
+	// This defines the flags accepted by this command. The key is the longname, the char property is the shortname,
+	// and summary and description is what's printed when the -h/--help flag is supplied.
 	public static readonly flags = {
 		language: Flags.string({
 			char: 'l',
@@ -35,34 +36,7 @@ export default class Add extends ScannerCommand {
 		})()
 	};
 
-	public readonly pathResolver: PathResolver;
-
-	public constructor(argv: string[], config: Config,
-					   inputValidatorFactory?: InputValidatorFactory,
-					   pathResolver?: PathResolver) {
-		if (typeof inputValidatorFactory === 'undefined') {
-			inputValidatorFactory = new RuleAddCommandInputValidatorFactory();
-		}
-		if (typeof pathResolver === 'undefined') {
-			pathResolver = new PathResolverImpl();
-		}
-		super(argv, config, inputValidatorFactory);
-		this.pathResolver = pathResolver;
-	}
-
-	async runInternal(inputs: Inputs): Promise<AnyJson> {
-		const language = inputs.language as string;
-		const paths = this.pathResolver.resolvePaths(inputs);
-
-		this.logger.trace(`Language: ${language}`);
-		this.logger.trace(`Rule path: ${JSON.stringify(paths)}`);
-
-		// Add to Custom Classpath registry
-		const manager = await Controller.createRulePathManager();
-		const classpathEntries = await manager.addPathsForLanguage(language, paths);
-
-		this.display.displayInfo(`Successfully added rules for ${language}.`);
-		this.display.displayInfo(`${classpathEntries.length} Path(s) added: ${classpathEntries.toString()}`);
-		return {success: true, language, path: classpathEntries};
+	protected createAction(logger: Logger, display: Display): Action {
+		return new RuleAddAction(logger, display, new PathResolverImpl());
 	}
 }
