+++ src/lib/ScannerCommand.ts
@@ -3,57 +3,63 @@ import {uxEvents, EVENTS} from './ScannerEvents';
 import {initContainer} from '../ioc.config';
 import {AnyJson} from '@salesforce/ts-types';
 import {Inputs} from '../types';
-import {Logger} from '@salesforce/core';
 import {Display, Displayable, UxDisplay} from "./Display";
 import {Bundle, getMessage} from "../MessageCatalog";
-import {Config} from "@oclif/core";
-import {InputValidatorFactory} from "./InputValidator";
+import {Logger} from "@salesforce/core";
 
 
+/**
+ * Abstract class to help build scanner commands.
+ *
+ * The responsibilities of this parent class is to:
+ *   * Parse the cli input flags
+ *   * To construct the Logger and Display instances
+ *   * To ask the subclass for the Action instance and invoke the validateInputs and run methods on it.
+ *
+ * The responsibilities of each subclass is to:
+ *   * Define the summary, description, examples, and flags for its command
+ *   * Construct the Action associated with its command with all of its Runtime dependencies
+ *
+ * Implementations of ScannerCommand should ideally have zero business logic in them since they should simply serve as
+ * runtime dependency injection points to their corresponding Action classes.
+ */
 export abstract class ScannerCommand extends SfCommand<AnyJson> implements Displayable {
-	protected logger: Logger;
-	protected display: Display;
-	private inputValidatorFactory: InputValidatorFactory;
-
-	protected constructor(argv: string[], config: Config, inputValidatorFactory: InputValidatorFactory) {
-		super(argv, config);
-		this.inputValidatorFactory = inputValidatorFactory;
-	}
-
 	protected async init(): Promise<void> {
 		await super.init();
 		initContainer();
 	}
 
+	protected abstract createAction(logger: Logger, display: Display) : Action;
+
 	public async run(): Promise<AnyJson> {
-		this.logger = await Logger.child(this.ctor.name);
-
 		const inputs: Inputs = (await this.parse(this.ctor)).flags;
-		this.display = new UxDisplay(this, this.spinner, inputs.verbose);
-		await this.inputValidatorFactory.createInputValidator(this.display).validate(inputs);
+		const logger: Logger = await Logger.child(this.ctor.name);
+		const display: Display = new UxDisplay(this, this.spinner, inputs.verbose);
 
-		this.display.displayWarning(getMessage(Bundle.Common, 'surveyRequestMessage'));
-		this.buildEventListeners();
+		display.displayWarning(getMessage(Bundle.Common, 'surveyRequestMessage'));
+		this.buildEventListeners(display);
 
-		return await this.runInternal(inputs);
+		const action: Action = this.createAction(logger, display);
+		await action.validateInputs(inputs);
+		return await action.run(inputs);
 	}
-
-	/**
-	 * Command's should implement this method to add their
-	 * working steps.
-	 */
-	abstract runInternal(inputs: Inputs): Promise<AnyJson>;
 
 	// TODO: We should consider refactoring away from events and instead inject the "Display" as a dependency
 	// into all of the classes that emit events
-	private buildEventListeners(): void {
-		uxEvents.on(EVENTS.INFO_ALWAYS, (msg: string) => this.display.displayInfo(msg));
-		uxEvents.on(EVENTS.INFO_VERBOSE, (msg: string) => this.display.displayVerboseInfo(msg));
-		uxEvents.on(EVENTS.WARNING_ALWAYS, (msg: string) => this.display.displayWarning(msg));
-		uxEvents.on(EVENTS.WARNING_VERBOSE, (msg: string) => this.display.displayVerboseWarning(msg));
-		uxEvents.on(EVENTS.START_SPINNER, (msg: string, status: string) => this.display.spinnerStart(msg, status));
-		uxEvents.on(EVENTS.UPDATE_SPINNER, (msg: string) => this.display.spinnerUpdate(msg));
-		uxEvents.on(EVENTS.WAIT_ON_SPINNER, (msg: string) => this.display.spinnerWait());
-		uxEvents.on(EVENTS.STOP_SPINNER, (msg: string) => this.display.spinnerStop(msg));
+	private buildEventListeners(display: Display): void {
+		uxEvents.on(EVENTS.INFO_ALWAYS, (msg: string) => display.displayInfo(msg));
+		uxEvents.on(EVENTS.INFO_VERBOSE, (msg: string) => display.displayVerboseInfo(msg));
+		uxEvents.on(EVENTS.WARNING_ALWAYS, (msg: string) => display.displayWarning(msg));
+		uxEvents.on(EVENTS.WARNING_VERBOSE, (msg: string) => display.displayVerboseWarning(msg));
+		uxEvents.on(EVENTS.START_SPINNER, (msg: string, status: string) => display.spinnerStart(msg, status));
+		uxEvents.on(EVENTS.UPDATE_SPINNER, (msg: string) => display.spinnerUpdate(msg));
+		uxEvents.on(EVENTS.WAIT_ON_SPINNER, (msg: string) => display.spinnerWait());
+		uxEvents.on(EVENTS.STOP_SPINNER, (msg: string) => display.spinnerStop(msg));
 	}
+}
+
+export interface Action {
+	validateInputs(inputs: Inputs): Promise<void>;
+
+	run(inputs: Inputs): Promise<AnyJson>;
 }
